Revert "Styling2"
