+++ client/src/App.css
@@ -13,7 +13,16 @@
   }
 }
 
-
+.App-header {
+  background-color: #282c34;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+}
 
 .App-link {
   color: #61dafb;
@@ -27,3 +36,36 @@
     transform: rotate(360deg);
   }
 }
+
+.homepage {
+  
+  color: #61dafb;
+  display: flex;
+  padding: 2% aqua;
+  background-color:bisque;
+  align-content: center;
+}
+#Avail_users {
+  display: inline-flex;
+  flex-flow: column nowrap;
+  align-items: stretch;
+  background-color: whitesmoke;
+  border: 3px dotted #fbae61;
+  border-spacing: 2px;
+  border-width: 5px;
+  width: 40%;
+  display: flex-start;
+}
+#Avail_pets {
+  display: inline-flex;
+  flex-flow: column nowrap;
+  align-items: stretch;
+  background-color: gray;
+  border: 3px dotted black;
+  border-spacing: 2px;
+  border-width: 5px;
+  width: 40%;
+  align-content: flex-end;
+  display: flex-start;
+}
+
