--- client/src/components/Footer/index.js
@@ -1,8 +1,0 @@
-import './Footer.css';
-
-const Footer = () =>
-  <div>
-      <h1>The Pet People</h1>
-  </div>
-
-  export default Footer;
