--- client/src/components/Header/Header.css
@@ -1,10 +1,0 @@
-.App-header {
-  background-color: #056519;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
