--- client/src/components/Header/index.js
@@ -1,8 +1,0 @@
-import './Header.css';
-
-const Header = () =>
-  <div>
-      <h1>The Pet People</h1>
-  </div>
-
-  export default Header;
