--- client/src/components/Homepage/Homepage.css
@@ -1,31 +1,0 @@
-/* .homepage {
-  
-  color: #61dafb;
-  display: flex;
-  padding: 2% aqua;
-  background-color:bisque;
-  align-content: center;
-} */
-/* #Avail_users {
-  display: flex;
-  flex-flow: column nowrap;
-  align-items: stretch;
-  background-color: whitesmoke;
-  border: 3px dotted #fbae61;
-  border-spacing: 2px;
-  border-width: 5px;
-  width: 40%;
-  display: flex-start; 
-} */
-/*#Avail_properties {
-  display: inline-flex;
-  flex-flow: column nowrap;
-  align-items: stretch;
-  background-color: gray;
-  border: 3px dotted black;
-  border-spacing: 2px;
-  border-width: 5px;
-  width: 40%;
-  align-content: flex-end;
-  display: flex-start;
-} */
