--- client/src/components/Homepage/index.js
@@ -1,8 +1,0 @@
-import './Homepage.css';
-
-const Homepage = () =>
-  <div>
-      <h1>The Pet People</h1>
-  </div>
-
-  export default Homepage;
