--- client/src/components/LoginForm/index.js
@@ -1,8 +1,0 @@
-import './LoginForm.css';
-
-const LoginForm = () =>
-  <div>
-      <h1>The Pet People</h1>
-  </div>
-
-  export default LoginForm;
