+++ client/src/components/Navbar.js
@@ -1,10 +1,10 @@
 import React, { useState } from "react";
 import { Link } from "react-router-dom";
 import { Navbar, Nav, Container, Modal, Tab } from "react-bootstrap";
-import SignUpForm from "../SignupForm/SignupForm";
-import LoginForm from "../LoginForm/LoginForm";
+import SignUpForm from "./SignupForm";
+import LoginForm from "./LoginForm";
 
-import Auth from "../../utils/auth";
+import Auth from "../utils/auth";
 
 const AppNavbar = () => {
   // set modal display state
