--- client/src/components/Navbar/index.js
@@ -1,9 +1,0 @@
-import { Nav } from 'react-bootstrap';
-import './Navbar.css';
-
-const Navbar = () =>
-  <div>
-      <h1>The Pet People</h1>
-  </div>
-
-  export default Navbar;
