--- client/src/components/Profile/index.js
@@ -1,8 +1,0 @@
-import './Profile.css';
-
-const Profile = () =>
-  <div>
-      <h1>The Pet People</h1>
-  </div>
-
-  export default Profile;
