+++ client/src/components/SignupForm.js
@@ -1,36 +1,21 @@
 import React, { useState } from "react";
 import { Form, Button, Alert } from "react-bootstrap";
-import { useMutation, useQuery } from "@apollo/client";
-import { UPDATE_USER } from "../../utils/mutations";
-import { GET_ME } from "../../utils/queries";
+import { useMutation } from "@apollo/client";
+import { ADD_USER } from "../utils/mutations";
+import Auth from "../utils/auth";
 
-import Auth from "../../utils/auth";
-
-const UpdateProfile = () => {
+const SignupForm = () => {
   // set initial form state
   const [userFormData, setUserFormData] = useState({
     username: "",
     email: "",
     password: "",
-    img: "",
-    location: "",
   });
   // set state for form validation
   const [validated] = useState(false);
   // set state for alert
   const [showAlert, setShowAlert] = useState(false);
-  const [updateUser] = useMutation(UPDATE_USER);
-  const { loading, data: userData } = useQuery(GET_ME);
-  // check if user is authenticated
-  const token = Auth.loggedIn() ? Auth.getToken() : null;
-
-  if (loading) {
-    return <h2>LOADING...</h2>;
-  }
-
-  if (!token) {
-    return <p>You must be logged in to update your profile.</p>;
-  }
+  const [addUser] = useMutation(ADD_USER);
 
   const handleInputChange = (event) => {
     const { name, value } = event.target;
@@ -43,26 +28,14 @@ const UpdateProfile = () => {
     // check if form has everything (as per react-bootstrap docs)
     const form = event.currentTarget;
     if (form.checkValidity() === false) {
-      event.persist();
+      event.preventDefault();
       event.stopPropagation();
     }
 
     try {
-      const username = userFormData.username;
-      const email = userFormData.email;
-      const img = userFormData.img;
-      const location = userFormData.location;
+      const { data } = await addUser({ variables: userFormData });
 
-      const { data } = await updateUser({
-        variables: {
-          username,
-          email,
-          img,
-          location,
-        },
-      }, {userData});
-
-      Auth.getProfile(data.token);
+      Auth.login(data.addUser.token);
     } catch (err) {
       console.error(err);
       setShowAlert(true);
@@ -72,21 +45,11 @@ const UpdateProfile = () => {
       username: "",
       email: "",
       password: "",
-      img: "",
-      location: "",
     });
-
-    event.persist();
-    // on submit send user to their profile
-    window.location.href = '/profile';
   };
 
   return (
     <>
-      <p>
-        Hey {userData?.me.username} What would you like to update about your
-        profile!
-      </p>
       {/* This is needed for the validation functionality above */}
       <Form noValidate validated={validated} onSubmit={handleFormSubmit}>
         {/* show alert if server response is bad */}
@@ -96,7 +59,7 @@ const UpdateProfile = () => {
           show={showAlert}
           variant="danger"
         >
-          Something went wrong with updating your profile!
+          Something went wrong with your signup!
         </Alert>
 
         <Form.Group className="mb-3">
@@ -107,7 +70,11 @@ const UpdateProfile = () => {
             name="username"
             onChange={handleInputChange}
             value={userFormData.username}
+            required
           />
+          <Form.Control.Feedback type="invalid">
+            Username is required!
+          </Form.Control.Feedback>
         </Form.Group>
 
         <Form.Group className="mb-3">
@@ -118,34 +85,38 @@ const UpdateProfile = () => {
             name="email"
             onChange={handleInputChange}
             value={userFormData.email}
+            required
           />
+          <Form.Control.Feedback type="invalid">
+            Email is required!
+          </Form.Control.Feedback>
         </Form.Group>
 
         <Form.Group className="mb-3">
-          <Form.Label htmlFor="img">Profile Image</Form.Label>
+          <Form.Label htmlFor="password">Password</Form.Label>
           <Form.Control
-            type="file"
-            placeholder="Only City and state"
-            name="image"
+            type="password"
+            placeholder="Your password"
+            name="password"
             onChange={handleInputChange}
-            value={userFormData.img}
+            value={userFormData.password}
+            required
           />
+          <Form.Control.Feedback type="invalid">
+            Password is required!
+          </Form.Control.Feedback>
         </Form.Group>
-
-        <Form.Group className="mb-3">
-          <Form.Label htmlFor="location">
-            Where are you looking to move to. (only city and state){" "}
-          </Form.Label>
-          <Form.Control
-            type="text"
-            placeholder="Only City and state"
-            name="location"
-            onChange={handleInputChange}
-            value={userFormData.location}
-          />
-        </Form.Group>
-
-        <Button type="submit" variant="success">
+        <Button
+          disabled={
+            !(
+              userFormData.username &&
+              userFormData.email &&
+              userFormData.password
+            )
+          }
+          type="submit"
+          variant="success"
+        >
           Submit
         </Button>
       </Form>
@@ -153,4 +124,4 @@ const UpdateProfile = () => {
   );
 };
 
-export default UpdateProfile;
+export default SignupForm;
