--- client/src/components/SignupForm/index.js
@@ -1,8 +1,0 @@
-import './SignupForm.css';
-
-const SignupForm = () =>
-  <div>
-      <h1>The Pet People</h1>
-  </div>
-
-  export default SignupForm;
