--- client/src/components/UpdateProfile/index.js
@@ -1,8 +1,0 @@
-import './UpdateProfile.css';
-
-const UpdateProfile = () =>
-  <div>
-      <h1>The Pet People</h1>
-  </div>
-
-  export default UpdateProfile;
