--- client/src/components/UsersList/index.js
@@ -1,8 +1,0 @@
-import './UsersList.css';
-
-const UsersList = () =>
-  <div>
-      <h1>The Pet People</h1>
-  </div>
-
-  export default UsersList;
